TST: separate out grouping-type tests
